refactor(runtime): split runtime registry module
